requirements.txt
djangorestframework
selenium
playwright
django-cors-headers
selenium
requests
python-dotenv
djangorestframework-simplejwt
python-decouple